twitter/functions/twitter_ingest_queue_get/requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-firestore==2.0.2
google-cloud-pubsub==1.5.0
elasticsearch==7.13.4
elasticsearch-dsl>=7.0.0,<8.0.0
